requirements.txt
# Lexmount SDK (从腾讯云 COS 安装)
lexmount>=0.3.1

# Browser automation
playwright>=1.40.0

# Environment variables
python-dotenv>=0.19.0

# HTTP client with SOCKS proxy support
httpx[socks]>=0.23.0
